BUG#38260801 generate MRS SDK for TypeScript by omission
There are a bunch of weird corner-cases regarding the target language
when generating the MRS SDK via the MySQL Shell CLI. Originally, the
idea is to first honor the language specified when calling the
mrs.dump.sdk_service_files() command, and if none is specified, use the
one specified in the sdk manifest file, if it is not a new installation.

However, the workflow must provide a clear path to handle situations
where the value of any of those parameters is invalid, not supported, or
simply unavailable (in which case there should be a sane default).

Additionally, the manifest file is also being generated with invalid
parameters, if for some reason, the workflow fails. This should
definitely not happen.

This patch introduces some changes to address all these scenarios and
ensures the TypeScript SDK can be generated by omission for new
installations.

Change-Id: I4f526e55caa961a4092debcc63d3fe695acdc022