BUG#38251258 simplify MRS SDK generation in the Shell CLI
Generating the MRS SDK for a given service in the MySQL Shell CLI
currently requires the service to be the "current" service in the
corresponding Shell session, or for a specific service id to be
specified. On one hand, the former requires an extra step, the latter is
not exactly user-friendly because it requires the service id to be
retrieved from the metadata schema beforehand.

This patch introduces some changes to improve this workflow, in the form
of an extra parameter which allows to identify the service using its
request path, which is more user friendly. Additionally, it fixes some
corner cases in the event that a service is not explicitly identified.
This can happen when updating an existing installation, when relying on
the "current" service or when there is only one service to work with.
Moreover, the SDK should not be generated when a service is not enabled.

Change-Id: I3cb28d0d32cb8ec689e2f58eff5d0d21e2c979e2