Fixing random failure in sql_editor cancel request test (BE).
The sql_editor.cancel_request test fails randomly for two reasons:

- Timeout may occur while wainting the cancellation responses, leading
  to a missing response error.
- Request ID mistmatch on a final response.

The first problem is now solved by increasing the wait timeout while
expecting the request cancellation responses.

The second problem was caused by the test not waiting for the successful
termination of the first request that was not being cancelled, leading
to an request id mistmatch on the test termination logic which sends a
request to delete a connection but was then getting the OK response
for a previous request.

Additionally, to help debugging, a ScopedCallback class was introduced
to ensure the entire output of a single tests is enclosed in the following
tags:

    "===== STARTING EXECUTION ====="
    "====== ENDING EXECUTION ====="

In the past, if a test failed, the closing tag was never printed,
so the boundary with the next test was never clear.

Change-Id: I0e15ee30b505808e315e6b310e22e3147f863e05