Adding support for python shell plugin debugging
With this patch, BE developers will be able to debug python plugins
in the shell context, by selecting the "Debug Shell Plugins" launch
profile.

What it will do, is run the shell in background mode, starting a
python debug server in the same port VSCode uses to debug python
and will have VSCode attached to it.

Now the devs will be able to debug python plugins normally by calling
the needed functions in the shell command prompt.

Change-Id: I07525e714ccb2c6ef9a42b3da98495c06698e2d8