BUG#38336041 revert breaking changes in the TypeScript MRS SDK
A recent patch that introduced a new linter configuration and addressed
newfound linting issues in the TypeScript SDK static code files ended up
making a bunch of breaking changes in the SDK API.

Ultimately, the SDK files should have their own convention (although
based on what's used project-wide), but since the SDK files are linked
under the frontend project (in order to be loaded by extension editors
and notebooks), the linter is also yielding the corresponding errors and
warnings.

This patch reverts all the changes in the SDK files and updates the
linter configuration to ignore them under the frontend project. The
actualy linting issues in the SDK files should be addressed as part of a
different patch.

Change-Id: I250b08a6e4e411c98e99b731d79a22220144ea95