BUG#38089763 BUG#38043458 handle async routine output inconsistencies
REST procedures associated with an Async Task are currently not capable
of producing result sets, only values for any existing OUT/INOUT
parameter. Consequently, the JSON document produced by the status report
contains the mapping of output parameters at the root level instead of
using the "resultSets" and "outParameters" fields common to the message
produced as a result of calling a "regular" procedure.

Additionally, REST functions with an associated async task should
produce an object with an additional "result" field when the task is
being manually checked via the "watch()" command.

This patch updates the MRS Python and TypeScript SDKs to account for all
these inconsistencies.

Change-Id: Iad8cca43afe43d4db4553b2b81be6c42113bf90f