BUG#37575319 fix output of mrs.getStatus() command
In an open Shell GUI DB connection TypeScript Notebook, the
mrs.getStatus() command returns a JSON object containing details about
each REST service, including the associated deployment URL.

When setting a different service as the current one for the existing
Shell session, the deployment URL specified as the "serviceURL" field in
the current service definition, updates to an incorrect value.

This patch introduces some changes to address this issue.

Change-Id: Ib5b25996add61cefb94c39aacca78e4bf32abf2b