BUG#38384637 BUG#38384773 remove WKT support from TS SDK
The MRS TypeScript SDK has a bunch of issues related to spatial data
types, particulary caused by the fact that it tries to support WKT for
upstream operations such as INSERTs and UPDATEs. The corresponding type
aliases are only half-baked when it comes to strictness in enforcing a
valid WKT format. This leads to more harm than good, so, just like in
the Python SDK, for the sake of simplicity, WKT should not be supported.

This patch introduces a few changes and removes WKT support from the
TypeScript SDK to address a couple of specific issues that been
identified which are caused by these inconsistences in the client-side
type aliases generated for spatial data types.

Change-Id: I3f50472751bbfa277d5a62784041d52021e02c9b