BUG#37767208 nested field attributes in MRS PY SDK
In the MRS Python SDK, when working with a local instance of an MRS
document, the document fields are accessed as attributes through the use
of a intermediate dataclass. However, this is not the case for nested
fields, where the access pattern is based on dictionary subscripts, just
like in the case of a regular dict.

This patch introduces some changes to address this inconsistency and
make sure that nested field properties can be accessed as nested class
attributes.

Change-Id: I9851bfe7f06ca63d78d397412e365fb096d8d3ca