WL#15899: handle lossy numbers in the MRS TS SDK
Values of the BIGINT and DECIMAL/NUMERIC MySQL data types are being
converted to a string in TypeScript. This is because, due to the nature
of IEEE 754 numbers, integers above 2^53-1 and fixed point decimals,
cannot be reliably represented in the language using such numeric
instances.

However, 64 bit integers can be safely represented using the BigInt
type, and the it is possible to determine wether the downstream numeric
values will lose precision or not.

This patch introduces some changes to allow applications some
flexibility as to how they can represent such numeric values when using
the TypeScript SDK. Numeric values for the above-mentioned data types
can be provided as "number"s or "bigint"s in the case of BIGINT, or
"numbers" or "string" in the case of DECIMAL/NUMERIC. Downstream, they
are converted to the most convenient type that ensures there is no
precision loss.

Change-Id: I050df4d4f1354e9268f7d3a150c8aab214e22dd6