BUG#38461981 handle lossy numbers in MRS for Python
Values of DECIMAL and NUMERIC data types in MySQL are currently being
converted to Python floats. Since those data types can accommodate fixed
point numeric values that lose precision when represented as floating
points, this is not ideal.

This patch introduces support for converting those values to a
decimal.Decimal instance. For upstream operations such as INSERTs and
UPDATEs, it is up to the user whether to use a decimal.Decimal instance
or a float. The same applies when creating a query filter. For the
results of downstream operations (READs), a flot is always used if the
value does not lose precision, otherwise a decimal.Decimal instance is
used.

Change-Id: I0322b6a177bdb3f00fb8d51b04316d476a55ff9e