BUG#37697639 change referential action on mrs_role table
Dropping a REST Service that contains hierarchical roles results in a
foreign key violation. This happens because the mrs_role table in the
REST Service metadata schema references itself via one of the foreign
key constraints, and due to its hierarchical nature, the rows must be
deleted in a specific order.

This patch introduces some changes to ensure that specific order is
enforced, namely, by changing the referencial action of the foreign key
constraint to "ON DELETE CASCADE".

Change-Id: I257e01c9288870baef80e62ae8f277d257cd3a18