BUG#38463211 fix MRS TS SDK broken call sites
A previous patch to address WL#15899 introduced a few breaking changes
in the function signatures of the MRS SDK TypeScript internal API. The
patch missed changes in some of the call sites of those functions,
leading to type-check errors in the generated SDK.

The issue is only reproducible for updatable REST Views and/or REST
Functions using the Async Task framework.

This patch introduces the changes to the call sites missed by the patch
that introduced the regression in the first place.

Change-Id: Ia7043d24a01476fd9f09da86f145a39d20423e4c