BUG#38558444 use number array for VECTOR in TS SDK
Instances of the VECTOR data type are currently expected to be
represented as "string"s in TypeScript applications. This is not ideal,
particularly for upstream data that is sent by the application to the
MRS backend when building query filters, inserting/updating documents or
calling routines, because there is no way to actually convert those
values back to a VECTOR.

On the other way, the MRS backend is capable of converting those values,
both for upstream and downstream, to and from an array of numbers.

This patch introduces some changes in the TypeScript MRS SDK to ensure
those values are expected to be represented as an array of numbers in
TypeScript applications.

Change-Id: I178794751604ffe4b40aa221d22069ebf24c240a