BUG#38576481 fix nested custom data type imports in PY SDK
MySQL data types such as VECTOR do not have a direct 1-1 mapping to any
primitive Python type. In those cases, a corresponding type alias is
created that accounts for the proper client side type constraints. These
type aliases must be imported by the generated SDK files.

Although this seems to be working fine for custom data types assigned to
top-level fields of a REST View, it is not the case for nested fields.

This patch introduces some changes to the MRS SDK code generator, to
ensure those type aliases are imported by the generated SDK files, when
required.

Change-Id: I106970c79708d8175998bd89eb53070af1db617d