BUG#38576785 fix PK name case in the MRS Python SDK
The document-level update() command in the Python SDK, requires the
actual name of the primary key fields in a REST View to ensure the
correct type mapping in the data class which represents said document.

In MRS, the raw field names are specified in lowerCamelCase, however, in
the scope of the Python SDK, these must be converted to snake_case to
better match the language convention. This is currently not happening
and the command will raise a runtime error whilst trying to get the
value of a field because the field name is not correct.

This patch introduces the changes required to ensure that primary key
field names in the scope of a REST document dataclass in the Python SDK
are using snake_case.

Change-Id: Ibdbdbc7639065241dc702ddc6cd675c995a35063