BUG#38593662 BUG#38593678 avoid nested field colisions
When creating a REST View in MRS, field names that map to foreign key
columns in a corresponding table are produced optimistically by first
using the referenced table name, and only if there already exists a
duplicate, proceed to append a suffix based on the name of the actual
foreign key column.

This logic is only followed if the REST View is created in the GUI, but
not when it is created using the MRS SQL flavour leading to weird
situations where a nested field can be ignored under specific
circumstances, such as having a colision between the REST View field
name, and the underlying referenced table name.

In order to address this, and not overcomplicate the process in the SQL
parser business logic, this patch introduces some changes to ensure the
field name is always unique, by always appending the column name suffix.
In turn, this ensures that the generated SQL is always deterministic.

Change-Id: I710285bff8cc19cf580bdacbb51acbb78fb6df8b