BUG#38623232 fix MSG certificate setup hyperlink
If the MSG self-signed certificate is not installed in the system,
and/or not trusted on macOS, the extension fails to start and displays
an error message that contains a broken hyperlink.

This patch fixes the hyperlink and ensures it directs users to the
appropriate documentation web page that describes how to manually
install the certificate.

Change-Id: I806efad62eda7b2c38aa6705c2cdab7fb9cd387b