Fixed problem with newer Node.js versions
The vite-plugin-monaco-editor node package uses a no longer supported version of rmDirSync. Warnings for that have been printed for a long time, but now that API was removed, leading to trouble when running the vite dev server. Another search and replace action was added to the build scripts to overcome that problem.
Additionally, a menu item was disabled that should not be enabled currently.

Change-Id: I406cfa7cc9fff1ab6575b1a48b8745bfc31cd59f