BUG#35420780 restore session state after configuring MRS
When the initial MRS configuration is done via the MySQL Shell using
either SQL or the Python API, a bunch of the operations that get
executed end up modifying the existing Shell session. In particular, the
default schema is changed multiple times in the SQL scripts executed by
the schema management plugin. This can be confusing to users.

This patch ensures the MRS configuration process happens in the scope of
a new Shell session, whilst keeping the state of the current global user
session.

Change-Id: I9f9235108daebc2717120928af206c58529c357f