BUG#38749096 make MRS auth params optional in notebook
When authenticating to a given REST service in a DB Notebook prompt
using the MySQL Shell GUI Extension for VS Code, authentication
parameters such as the username and the auth app name should be
optional. This is because the user is always displayed a dialog to enter
additional details, like the password, manually.

This patch introduces some changes to ensure these parameters are
actually optional in the scope of a DB Notebook prompt.

Change-Id: Id78714cfb436e958b464de84be53bdca98a0f51f