Fixing initialization error in WSL2
Part of the initialization code depends on the execution context,
determining the code is running in the VSCode extension is critical
to avoid specific functionality not available there.

This identification was NOT working when running in WSL, as it is
determined by the environment, however, the environment differs when
the extension is being launched vs when it is already executing, none
of the used criterias was correctly rutning ON the inExtension flag.

This patch fixes this by considering the environment when the extension
is being bootstrapped by VSCode Remote.

Change-Id: I17bd468f79acaf4e29353d8cca2c0e1e26734e0b